pyproject.toml
[tool.poetry]
name = "myVenv"
version = "0.1.0"
description = ""
authors = ["asd24684782 <asd24684782@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.0.1"
python-dotenv = "^0.21.0"
pydantic = "^1.10.2"
psycopg2 = "^2.9.9"

[tool.poetry.scripts]
discord = "ting-tuan-zi-discordBot.src.main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
